clean up, standardization and dynamic link for audio
