+++ core/api.py
@@ -8,31 +8,31 @@
 """
 
 import markdown
-from flask import Flask
 from flask import render_template
 from flask import redirect
 from flask import url_for
 from flask import Markup
+from flask import Flask
 
-from flawcode import app
-from flawcode.helpers import show_notes
-from flawcode.helpers import directly_linked_old
+from core.helpers import show_notes
+from core.helpers import get_last_4episode_num
+from core.settings import EPISODE_COUNT
 
-
-EPISODE_COUNT = 4
+app = Flask(__name__)
+app.config.from_object('core.settings')
 
 
 @app.route('/')
 def index():
-    return redirect(url_for('shows', ep_no='4'))
+    return redirect(url_for('shows', ep_no=str(EPISODE_COUNT)))
 
 
 @app.route('/episode/show/<ep_no>')
 def shows(ep_no):
     content = Markup(markdown.markdown(show_notes(ep_no)))
     return render_template("shows.html", name=ep_no,
-                           episodes=directly_linked_old(EPISODE_COUNT),
-                           **locals())
+                           episodes=get_last_4episode_num(EPISODE_COUNT),
+                           content=content)
 
 
 @app.errorhandler(404)
