+++ core/settings.py
@@ -1,0 +1,3 @@
+DEBUG = True
+SECRET_KEY = '6f75ba0f786636c79312e7e569278ccde6f379c8'
+EPISODE_COUNT = 4
