+++ core/show_notes/1.md
