+++ core/show_notes/2.md
