+++ core/show_notes/4.md
