--- flawcode/flawcode/__init__.py
@@ -1,6 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-app.config.from_object('flawcode.settings')
-
-import flawcode.views
