--- flawcode/flawcode/settings.py
@@ -1,2 +1,0 @@
-DEBUG = True
-SECRET_KEY = '6f75ba0f786636c79312e7e569278ccde6f379c8'
