--- flawcode/runserver.py
@@ -1,3 +1,0 @@
-from flawcode import app
-
-app.run()
