+++ runserver.py
@@ -1,0 +1,4 @@
+from core import api
+
+if __name__ == '__main__':
+    api.app.run()
